world: extract schema helpers
